requirements.txt
Flask
Flask-RESTful
google-cloud-storage
pyopenssl